+++ sdk/textanalytics/Azure.AI.TextAnalytics/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Release History
 
 ## 5.1.0-beta.7 (Unreleased)
-
+### Breaking changes
+- The client defaults to the latest supported service version, which currently is `3.1-preview.5`.
+- Renamed type `TextElementsV8` to `TextElementV8` in model `StringIndexType`.
 
 ## 5.1.0-beta.6 (2021-04-06)
 ### New features